pyproject.toml
[tool.poetry]
name = "RAG-for-production"
version = "0.1.0"
description = ""
authors = ["Annthomy Gilles <professional@willofv.mozmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pre-commit = "^3.6.0"
ray = { extras = ["default"], version = "^2.9.0" }
pandas = "^2.1.4"
pyarrow = "^14.0.2"
beautifulsoup4 = "^4.12.2"
ruff = "^0.1.9"
tqdm = "^4.66.1"
langchain = "^0.0.353"
python-dotenv = "^1.0.0"
pgvector = "^0.2.4"
psycopg = { extras = ["binary"], version = "^3.1.16" }
openai = "^1.6.1"
tiktoken = "^0.5.2"
pillow = "^10.2.0"
loguru = "^0.7.2"
chromadb = "^0.4.22"
arxiv = "^2.1.0"
pypdf = "^3.17.4"
ratelimit = "^2.2.1"
types-requests = "^2.31.0.20240106"
mypy = "^1.8.0"
requests = "^2.31.0"
sentence-transformers = "^2.2.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
